Transformation module, init logs, some cleanup: - transformation related functions moved to own handler with required imports - added console logs for module inits just in case - ran format documents for scritps
