+++ src/clients/merge/common/dataHandler.js
@@ -1,16 +1,16 @@
 
 
 export {
-    deleteFromTransformed,
-    flipEditMode,
-    getClientName,
-    getEditMode,
-    getKeys,
-    getTransformed,
-    getUseProfileType,
-    init,
-    resetTransformed,
-    updateTransformed
+  deleteFromTransformed,
+  flipEditMode,
+  getClientName,
+  getEditMode,
+  getKeys,
+  getTransformed,
+  getUseProfileType,
+  init,
+  resetTransformed,
+  updateTransformed
 };
 
 //for transformed data see init and initTransformed functions
@@ -27,19 +27,20 @@ const keys = {
 // Exported
 //-----------------------------------------------------------------------------
 
-function init(data = {}){
-    const { canUseProfileType = true, transformedDefaultOptions = undefined, client = 'muuntaja' } = data;
+function init(data = {}) {
+  console.log('data module OK');
+  const { canUseProfileType = true, transformedDefaultOptions = undefined, client = 'muuntaja' } = data;
 
-    transformed = Object.create(initTransformed());
-    editMode = false;
-    useProfileType = canUseProfileType;
-    clientName = client;
+  transformed = Object.create(initTransformed());
+  editMode = false;
+  useProfileType = canUseProfileType;
+  clientName = client;
 
-    //override options
-    if (transformedDefaultOptions) {
-        transformed.options = transformedDefaultOptions;
-        overrideOptions = transformedDefaultOptions;
-    }
+  //override options
+  if (transformedDefaultOptions) {
+    transformed.options = transformedDefaultOptions;
+    overrideOptions = transformedDefaultOptions;
+  }
 }
 
 //common transformed edit functions
@@ -47,23 +48,23 @@ function getTransformed() { return transformed; }
 function updateTransformed(updateData) { transformed.update(updateData); }
 function deleteFromTransformed(propertyPathInString) { transformed.deleteProperty(propertyPathInString); }
 function resetTransformed() {
-    transformed = Object.create(initTransformed());
-    //override options
-    if (overrideOptions) {
-        transformed.options = overrideOptions;
-    }
+  transformed = Object.create(initTransformed());
+  //override options
+  if (overrideOptions) {
+    transformed.options = overrideOptions;
+  }
 }
 
 //common keys
-function getKeys(){return keys;}
+function getKeys() { return keys; }
 
 //edit mode (edit mode is globally available through window but lets try to avoid direct mutations)
-function flipEditMode(){editMode = !editMode;}
-function getEditMode(){return editMode;}
+function flipEditMode() { editMode = !editMode; }
+function getEditMode() { return editMode; }
 
 //other get
-function getUseProfileType(){return useProfileType;}
-function getClientName(){return clientName;}
+function getUseProfileType() { return useProfileType; }
+function getClientName() { return clientName; }
 
 //-----------------------------------------------------------------------------
 // Private
@@ -75,38 +76,38 @@ function getClientName(){return clientName;}
  * @returns {object} unmodified transformed object
  */
 function initTransformed() {
-    return {
-      options: {},
-      source: null,
-      base: null,
-      exclude: {},
-      replace: {},
-      insert: [],
-      //result: null,
-      //stored: null,
+  return {
+    options: {},
+    source: null,
+    base: null,
+    exclude: {},
+    replace: {},
+    insert: [],
+    //result: null,
+    //stored: null,
 
-      update(newData) {
-        Object.assign(this, newData);
-      },
-      deleteProperty(propertyPathInString) {
-        //simple object field
-        if (propertyPathInString in this) {
-          delete this[propertyPathInString];
+    update(newData) {
+      Object.assign(this, newData);
+    },
+    deleteProperty(propertyPathInString) {
+      //simple object field
+      if (propertyPathInString in this) {
+        delete this[propertyPathInString];
+        return;
+      }
+
+      //nested property
+      const propertiesInPath = propertyPathInString.split('.');
+      let currentObj = this;
+      for (const property of propertiesInPath.slice(0, -1)) {
+        if (!(property in currentObj)) {
+          //console.log(`Nested property '${propertyPathInString}' does not exist.`);
           return;
         }
-
-        //nested property
-        const propertiesInPath = propertyPathInString.split('.');
-        let currentObj = this;
-        for (const property of propertiesInPath.slice(0, -1)) {
-          if (!(property in currentObj)) {
-            //console.log(`Nested property '${propertyPathInString}' does not exist.`);
-            return;
-          }
-          currentObj = currentObj[property];
-        }
-        delete currentObj[propertiesInPath[propertiesInPath.length - 1]];
+        currentObj = currentObj[property];
       }
-    };
+      delete currentObj[propertiesInPath[propertiesInPath.length - 1]];
+    }
+  };
 }
