+++ src/clients/merge/common/urlHandler.js
@@ -7,7 +7,8 @@ export {
 //-----------------------------------------------------------------------------
 // Exported
 //-----------------------------------------------------------------------------
-function init(){}
+function init() { console.log('url module OK'); }
+
 /**
  * Parse data from url parameters
  * get data from url parameters (has defaults) and sets id to source and bas
@@ -24,18 +25,18 @@ function parseUrlParameters() {
 
     //only read additional parameters if prompted by client to do so
     if (dataModule.getUseProfileType()) {
-      const type = urlParams.get('type') || 'p2e';
-      const profile = urlParams.get('profile') || 'DEFAULT';
+        const type = urlParams.get('type') || 'p2e';
+        const profile = urlParams.get('profile') || 'DEFAULT';
 
-      document.querySelector('#type-options [name=\'type\']').value = type;
-      document.querySelector('#profile-options [name=\'profile\']').value = profile;
+        document.querySelector('#type-options [name=\'type\']').value = type;
+        document.querySelector('#profile-options [name=\'profile\']').value = profile;
 
-      dataModule.updateTransformed({
-        options:{
-          type:type,
-          profile: profile
-        }
-      });
+        dataModule.updateTransformed({
+            options: {
+                type: type,
+                profile: profile
+            }
+        });
     }
 }
 
@@ -54,10 +55,10 @@ function updateUrlParameters(transformed) {
     const { options, source, base } = transformed;
 
     const urlParamConfigs = [
-      {id:"type", defaultPosition: 1, value:options?.type, isActive: dataModule.getUseProfileType()},
-      {id:"profile", defaultPosition: 2,value:options?.profile, isActive: dataModule.getUseProfileType()},
-      {id:"baseId", defaultPosition: 4, value:base?.ID, isActive: true},
-      {id:"sourceId", defaultPosition: 3,value:source?.ID, isActive: true}
+        { id: "type", defaultPosition: 1, value: options?.type, isActive: dataModule.getUseProfileType() },
+        { id: "profile", defaultPosition: 2, value: options?.profile, isActive: dataModule.getUseProfileType() },
+        { id: "baseId", defaultPosition: 4, value: base?.ID, isActive: true },
+        { id: "sourceId", defaultPosition: 3, value: source?.ID, isActive: true }
     ];
     //filer not needed ones out, (those with not value or not active are left out)
     const filteredParamConfigs = urlParamConfigs.filter(obj => (obj.isActive && obj.value));
@@ -71,7 +72,7 @@ function updateUrlParameters(transformed) {
     //handle state
     window.history.replaceState({}, '', decodeURIComponent(`${window.location.pathname}?${urlParams}`));
     if (window.location.search === '') {
-      window.history.replaceState({}, '', `/${dataModule.getClientName()}/`);
+        window.history.replaceState({}, '', `/${dataModule.getClientName()}/`);
     }
 }
